@@ -1,13 +1,20 @@
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
 [project]
 name = "vonbot"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.12"
+requires-python = "==3.14"
 dependencies = [
     "pandas>=2.3.3",
     "polars>=1.35.2",
     "pymssql>=2.3.9",
     "python-dotenv>=1.2.1",
-    "streamlit>=1.51.0",
+    "streamlit>=1.40.0",
 ]
+
+[tool.setuptools.packages.find]
+include = ["logic", "models", "database", "services"]
