@@ -5,14 +5,15 @@ logger = logging.getLogger(__name__)
 
 DB_PATH = "admin_session.db"
 
+
 def init_local_db():
     """Initialize the local SQLite database for admin session data."""
     try:
         conn = sqlite3.connect(DB_PATH)
         cursor = conn.cursor()
-        
+
         # Table for storing temporary student lists
-        cursor.execute('''
+        cursor.execute("""
             CREATE TABLE IF NOT EXISTS session_students (
                 id INTEGER PRIMARY KEY AUTOINCREMENT,
                 student_id TEXT,
@@ -20,10 +21,10 @@ def init_local_db():
                 major TEXT,
                 selected BOOLEAN DEFAULT 1
             )
-        ''')
-        
+        """)
+
         # Table for storing course recommendations
-        cursor.execute('''
+        cursor.execute("""
             CREATE TABLE IF NOT EXISTS course_recommendations (
                 id INTEGER PRIMARY KEY AUTOINCREMENT,
                 student_id TEXT,
@@ -32,13 +33,14 @@ def init_local_db():
                 priority INTEGER,
                 status TEXT DEFAULT 'pending'
             )
-        ''')
-        
+        """)
+
         conn.commit()
         conn.close()
         logger.info("Local database initialized.")
     except Exception as e:
         logger.error(f"Failed to initialize local DB: {e}")
+
 
 def get_local_connection():
     return sqlite3.connect(DB_PATH)
